+++ pkg/objectupdate/sched/sched.go
@@ -27,7 +27,6 @@ import (
 	k8swgobjupdate "github.com/k8stopologyawareschedwg/deployer/pkg/objectupdate"
 	k8swgschedupdate "github.com/k8stopologyawareschedwg/deployer/pkg/objectupdate/sched"
 
-	nropv1 "github.com/openshift-kni/numaresources-operator/api/v1"
 	"github.com/openshift-kni/numaresources-operator/pkg/hash"
 	schedstate "github.com/openshift-kni/numaresources-operator/pkg/numaresourcesscheduler/objectstate/sched"
 )
@@ -38,15 +37,6 @@ const (
 	MainContainerName = "secondary-scheduler"
 )
 
-const (
-	PFPStatusDumpEnvVar = "PFP_STATUS_DUMP"
-
-	PFPStatusDir = "/run/pfpstatus"
-)
-
-// TODO: we should inject also the mount point. As it is now, the information is split between the manifest
-// and the updating logic, causing unnecessary friction. This code needs to know too much what's in the manifest.
-
 func DeploymentImageSettings(dp *appsv1.Deployment, userImageSpec string) {
 	cnt := k8swgobjupdate.FindContainerByName(dp.Spec.Template.Spec.Containers, MainContainerName)
 	if cnt == nil {
@@ -55,45 +45,6 @@ func DeploymentImageSettings(dp *appsv1.Deployment, userImageSpec string) {
 	}
 	cnt.Image = userImageSpec
 	klog.V(3).InfoS("Scheduler image", "reason", "user-provided", "pullSpec", userImageSpec)
-}
-
-func DeploymentEnvVarSettings(dp *appsv1.Deployment, spec nropv1.NUMAResourcesSchedulerSpec) {
-	cnt := k8swgobjupdate.FindContainerByName(dp.Spec.Template.Spec.Containers, MainContainerName)
-	if cnt == nil {
-		klog.ErrorS(nil, "cannot find container", "name", MainContainerName)
-		return
-	}
-
-	cacheResyncDebug := *spec.CacheResyncDebug
-	if cacheResyncDebug == nropv1.CacheResyncDebugDumpJSONFile {
-		setContainerEnvVar(cnt, PFPStatusDumpEnvVar, PFPStatusDir)
-	} else {
-		deleteContainerEnvVar(cnt, PFPStatusDumpEnvVar)
-	}
-}
-
-func setContainerEnvVar(cnt *corev1.Container, name, value string) {
-	if env := FindEnvVarByName(cnt.Env, name); env != nil {
-		klog.V(2).InfoS("overriding existing environment variable", "name", name, "oldValue", env.Value, "newValue", value)
-		env.Value = value
-		return
-	}
-
-	cnt.Env = append(cnt.Env, corev1.EnvVar{
-		Name:  name,
-		Value: value,
-	})
-}
-
-func deleteContainerEnvVar(cnt *corev1.Container, name string) {
-	var envs []corev1.EnvVar
-	for _, env := range cnt.Env {
-		if env.Name == name {
-			continue
-		}
-		envs = append(envs, env)
-	}
-	cnt.Env = envs
 }
 
 func FindEnvVarByName(envs []corev1.EnvVar, name string) *corev1.EnvVar {
